fix some server types/client types
